Generate prefixed Action Gateway API clients
