+++ src/dots/kiota-lock.json
@@ -1,8 +1,8 @@
 {
-  "descriptionHash": "8320CB83CF72F3CDB09AEE67C16F91A40A6EA7BF48168FC396601A700F6920CEF44380255B3FDF17EA41746788CD093D83EF8ED5E98B181DB46296344F8AA12D",
+  "descriptionHash": "8C210F6E7D540499F1C7482C2B55618EB8733F26159BBE3CCE21CB612869C08DE2F279FCF229F45498F3DA604EAD82171B952D8B2EDC47AADFE090C946A9D6AF",
   "descriptionLocation": "../../DigitalOcean-public.v2.yaml",
   "lockFileVersion": "1.0.0",
-  "kiotaVersion": "1.31.1",
+  "kiotaVersion": "1.34.1",
   "clientClassName": "DigitalOceanClient",
   "typeAccessModifier": "Public",
   "clientNamespaceName": "ApiSdk",
@@ -30,5 +30,6 @@
   ],
   "includePatterns": [],
   "excludePatterns": [],
-  "disabledValidationRules": []
+  "disabledValidationRules": [],
+  "allowedExternalOrigins": []
 }
