+++ src/dots/v2/actionGateway/connections/item/index.ts
@@ -1,0 +1,121 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createDelete_connection_responseFromDiscriminatorValue, createErrorEscapedFromDiscriminatorValue, createGet_connection_responseFromDiscriminatorValue, createUpdate_connection_parameters_responseFromDiscriminatorValue, serializeUpdate_connection_parameters_request, serializeUpdate_connection_parameters_response, type Delete_connection_response, type ErrorEscaped, type Get_connection_response, type Update_connection_parameters_request, type Update_connection_parameters_response } from '../../../../models/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/connections/{id}
+ */
+export interface ConnectionsItemRequestBuilder extends BaseRequestBuilder<ConnectionsItemRequestBuilder> {
+    /**
+     * Revokes and deletes an OAuth connection owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Delete_connection_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 404 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     delete(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Delete_connection_response | undefined>;
+    /**
+     * Retrieves an OAuth connection owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Get_connection_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 404 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Get_connection_response | undefined>;
+    /**
+     * Updates non-sensitive connection parameters for an OAuth connection.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Update_connection_parameters_response>}
+     * @throws {ErrorEscaped} error when the service returns a 400 status code
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 404 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     patch(body: Update_connection_parameters_request, requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Update_connection_parameters_response | undefined>;
+    /**
+     * Revokes and deletes an OAuth connection owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toDeleteRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+    /**
+     * Retrieves an OAuth connection owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+    /**
+     * Updates non-sensitive connection parameters for an OAuth connection.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toPatchRequestInformation(body: Update_connection_parameters_request, requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const ConnectionsItemRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/connections/{id}";
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const ConnectionsItemRequestBuilderRequestsMetadata: RequestsMetadata = {
+    delete: {
+        uriTemplate: ConnectionsItemRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createDelete_connection_responseFromDiscriminatorValue,
+    },
+    get: {
+        uriTemplate: ConnectionsItemRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createGet_connection_responseFromDiscriminatorValue,
+    },
+    patch: {
+        uriTemplate: ConnectionsItemRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            400: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createUpdate_connection_parameters_responseFromDiscriminatorValue,
+        requestBodyContentType: "application/json",
+        requestBodySerializer: serializeUpdate_connection_parameters_request,
+        requestInformationContentSetMethod: "setContentFromParsable",
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
