+++ src/dots/v2/actionGateway/index.ts
@@ -1,0 +1,72 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { ConnectionsRequestBuilderNavigationMetadata, ConnectionsRequestBuilderRequestsMetadata, type ConnectionsRequestBuilder } from './connections/index.js';
+// @ts-ignore
+import { SessionsRequestBuilderNavigationMetadata, SessionsRequestBuilderRequestsMetadata, type SessionsRequestBuilder } from './sessions/index.js';
+// @ts-ignore
+import { ToolbeltsRequestBuilderNavigationMetadata, ToolbeltsRequestBuilderRequestsMetadata, type ToolbeltsRequestBuilder } from './toolbelts/index.js';
+// @ts-ignore
+import { ToolsRequestBuilderNavigationMetadata, ToolsRequestBuilderRequestsMetadata, type ToolsRequestBuilder } from './tools/index.js';
+// @ts-ignore
+import { type UsersRequestBuilder, UsersRequestBuilderNavigationMetadata, UsersRequestBuilderRequestsMetadata } from './users/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway
+ */
+export interface ActionGatewayRequestBuilder extends BaseRequestBuilder<ActionGatewayRequestBuilder> {
+    /**
+     * The connections property
+     */
+    get connections(): ConnectionsRequestBuilder;
+    /**
+     * The sessions property
+     */
+    get sessions(): SessionsRequestBuilder;
+    /**
+     * The toolbelts property
+     */
+    get toolbelts(): ToolbeltsRequestBuilder;
+    /**
+     * The tools property
+     */
+    get tools(): ToolsRequestBuilder;
+    /**
+     * The users property
+     */
+    get users(): UsersRequestBuilder;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const ActionGatewayRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway";
+/**
+ * Metadata for all the navigation properties in the request builder.
+ */
+export const ActionGatewayRequestBuilderNavigationMetadata: Record<Exclude<keyof ActionGatewayRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
+    connections: {
+        requestsMetadata: ConnectionsRequestBuilderRequestsMetadata,
+        navigationMetadata: ConnectionsRequestBuilderNavigationMetadata,
+    },
+    sessions: {
+        requestsMetadata: SessionsRequestBuilderRequestsMetadata,
+        navigationMetadata: SessionsRequestBuilderNavigationMetadata,
+    },
+    toolbelts: {
+        requestsMetadata: ToolbeltsRequestBuilderRequestsMetadata,
+        navigationMetadata: ToolbeltsRequestBuilderNavigationMetadata,
+    },
+    tools: {
+        requestsMetadata: ToolsRequestBuilderRequestsMetadata,
+        navigationMetadata: ToolsRequestBuilderNavigationMetadata,
+    },
+    users: {
+        requestsMetadata: UsersRequestBuilderRequestsMetadata,
+        navigationMetadata: UsersRequestBuilderNavigationMetadata,
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
