+++ src/dots/v2/actionGateway/sessions/index.ts
@@ -1,0 +1,129 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createCreate_session_responseFromDiscriminatorValue, createErrorEscapedFromDiscriminatorValue, createList_sessions_responseFromDiscriminatorValue, serializeCreate_session_request, serializeCreate_session_response, type Create_session_request, type Create_session_response, type ErrorEscaped, type List_sessions_response } from '../../../models/index.js';
+// @ts-ignore
+import { type WithSession_urnItemRequestBuilder, WithSession_urnItemRequestBuilderRequestsMetadata } from './item/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/sessions
+ */
+export interface SessionsRequestBuilder extends BaseRequestBuilder<SessionsRequestBuilder> {
+    /**
+     * Gets an item from the ApiSdk.v2.actionGateway.sessions.item collection
+     * @param session_urn The URL-encoded managed agents session URN.
+     * @returns {WithSession_urnItemRequestBuilder}
+     */
+     bySession_urn(session_urn: string) : WithSession_urnItemRequestBuilder;
+    /**
+     * Lists Action Gateway sessions owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<List_sessions_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<SessionsRequestBuilderGetQueryParameters> | undefined) : Promise<List_sessions_response | undefined>;
+    /**
+     * Creates a session with a tool selection, invocation policy, and optional direct-tool preload configuration.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Create_session_response>}
+     * @throws {ErrorEscaped} error when the service returns a 400 status code
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     post(body: Create_session_request, requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Create_session_response | undefined>;
+    /**
+     * Lists Action Gateway sessions owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<SessionsRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
+    /**
+     * Creates a session with a tool selection, invocation policy, and optional direct-tool preload configuration.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toPostRequestInformation(body: Create_session_request, requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+}
+/**
+ * Lists Action Gateway sessions owned by the authenticated team.
+ */
+export interface SessionsRequestBuilderGetQueryParameters {
+    /**
+     * Filter sessions by actor identifier.
+     */
+    endUserId?: string;
+    /**
+     * Which 'page' of paginated results to return.
+     */
+    page?: number;
+    /**
+     * Number of items returned per page
+     */
+    perPage?: number;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const SessionsRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/sessions{?end_user_id*,page*,per_page*}";
+/**
+ * Mapper for query parameters from symbol name to serialization name represented as a constant.
+ */
+const SessionsRequestBuilderGetQueryParametersMapper: Record<string, string> = {
+    "endUserId": "end_user_id",
+    "perPage": "per_page",
+};
+/**
+ * Metadata for all the navigation properties in the request builder.
+ */
+export const SessionsRequestBuilderNavigationMetadata: Record<Exclude<keyof SessionsRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
+    bySession_urn: {
+        requestsMetadata: WithSession_urnItemRequestBuilderRequestsMetadata,
+        pathParametersMappings: ["session_urn"],
+    },
+};
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const SessionsRequestBuilderRequestsMetadata: RequestsMetadata = {
+    get: {
+        uriTemplate: SessionsRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createList_sessions_responseFromDiscriminatorValue,
+        queryParametersMapper: SessionsRequestBuilderGetQueryParametersMapper,
+    },
+    post: {
+        uriTemplate: SessionsRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            400: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createCreate_session_responseFromDiscriminatorValue,
+        requestBodyContentType: "application/json",
+        requestBodySerializer: serializeCreate_session_request,
+        requestInformationContentSetMethod: "setContentFromParsable",
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
