+++ src/dots/v2/actionGateway/toolbelts/index.ts
@@ -1,0 +1,137 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createErrorEscapedFromDiscriminatorValue, createToolbelt_responseFromDiscriminatorValue, createToolbelts_responseFromDiscriminatorValue, serializeToolbelt_create, serializeToolbelt_response, type ErrorEscaped, type Toolbelt_create, type Toolbelt_response, type Toolbelts_response } from '../../../models/index.js';
+// @ts-ignore
+import { type WithNameItemRequestBuilder, WithNameItemRequestBuilderNavigationMetadata, WithNameItemRequestBuilderRequestsMetadata } from './item/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+export type GetStatusQueryParameterType = (typeof GetStatusQueryParameterTypeObject)[keyof typeof GetStatusQueryParameterTypeObject];
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/toolbelts
+ */
+export interface ToolbeltsRequestBuilder extends BaseRequestBuilder<ToolbeltsRequestBuilder> {
+    /**
+     * Gets an item from the ApiSdk.v2.actionGateway.toolbelts.item collection
+     * @param name The natural key identifying the toolbelt.
+     * @returns {WithNameItemRequestBuilder}
+     */
+     byName(name: string) : WithNameItemRequestBuilder;
+    /**
+     * Lists the latest version of each toolbelt owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Toolbelts_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<ToolbeltsRequestBuilderGetQueryParameters> | undefined) : Promise<Toolbelts_response | undefined>;
+    /**
+     * Creates a versioned collection of provider-qualified Action Gateway tool names.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Toolbelt_response>}
+     * @throws {ErrorEscaped} error when the service returns a 400 status code
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 409 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     post(body: Toolbelt_create, requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Toolbelt_response | undefined>;
+    /**
+     * Lists the latest version of each toolbelt owned by the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<ToolbeltsRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
+    /**
+     * Creates a versioned collection of provider-qualified Action Gateway tool names.
+     * @param body The request body
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toPostRequestInformation(body: Toolbelt_create, requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+}
+/**
+ * Lists the latest version of each toolbelt owned by the authenticated team.
+ */
+export interface ToolbeltsRequestBuilderGetQueryParameters {
+    /**
+     * Which 'page' of paginated results to return.
+     */
+    page?: number;
+    /**
+     * Number of items returned per page
+     */
+    perPage?: number;
+    /**
+     * Filter toolbelts by status.
+     */
+    status?: GetStatusQueryParameterType;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const ToolbeltsRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/toolbelts{?page*,per_page*,status*}";
+export const GetStatusQueryParameterTypeObject = {
+    Active: "active",
+    Deprecated: "deprecated",
+    All: "all",
+} as const;
+/**
+ * Mapper for query parameters from symbol name to serialization name represented as a constant.
+ */
+const ToolbeltsRequestBuilderGetQueryParametersMapper: Record<string, string> = {
+    "perPage": "per_page",
+};
+/**
+ * Metadata for all the navigation properties in the request builder.
+ */
+export const ToolbeltsRequestBuilderNavigationMetadata: Record<Exclude<keyof ToolbeltsRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
+    byName: {
+        requestsMetadata: WithNameItemRequestBuilderRequestsMetadata,
+        navigationMetadata: WithNameItemRequestBuilderNavigationMetadata,
+        pathParametersMappings: ["name"],
+    },
+};
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const ToolbeltsRequestBuilderRequestsMetadata: RequestsMetadata = {
+    get: {
+        uriTemplate: ToolbeltsRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createToolbelts_responseFromDiscriminatorValue,
+        queryParametersMapper: ToolbeltsRequestBuilderGetQueryParametersMapper,
+    },
+    post: {
+        uriTemplate: ToolbeltsRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            400: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            409: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createToolbelt_responseFromDiscriminatorValue,
+        requestBodyContentType: "application/json",
+        requestBodySerializer: serializeToolbelt_create,
+        requestInformationContentSetMethod: "setContentFromParsable",
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
