+++ src/dots/v2/actionGateway/tools/index.ts
@@ -1,0 +1,112 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createErrorEscapedFromDiscriminatorValue, createList_tools_responseFromDiscriminatorValue, type ErrorEscaped, type List_tools_response } from '../../../models/index.js';
+// @ts-ignore
+import { type WithNameItemRequestBuilder, WithNameItemRequestBuilderNavigationMetadata } from './item/index.js';
+// @ts-ignore
+import { ProvidersRequestBuilderRequestsMetadata, type ProvidersRequestBuilder } from './providers/index.js';
+// @ts-ignore
+import { ToolkitsRequestBuilderRequestsMetadata, type ToolkitsRequestBuilder } from './toolkits/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/tools
+ */
+export interface ToolsRequestBuilder extends BaseRequestBuilder<ToolsRequestBuilder> {
+    /**
+     * The providers property
+     */
+    get providers(): ProvidersRequestBuilder;
+    /**
+     * The toolkits property
+     */
+    get toolkits(): ToolkitsRequestBuilder;
+    /**
+     * Gets an item from the ApiSdk.v2.actionGateway.tools.item collection
+     * @param name The provider-qualified tool name.
+     * @returns {WithNameItemRequestBuilder}
+     */
+     byName(name: string) : WithNameItemRequestBuilder;
+    /**
+     * Lists active Action Gateway tools visible to the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<List_tools_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<ToolsRequestBuilderGetQueryParameters> | undefined) : Promise<List_tools_response | undefined>;
+    /**
+     * Lists active Action Gateway tools visible to the authenticated team.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<ToolsRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
+}
+/**
+ * Lists active Action Gateway tools visible to the authenticated team.
+ */
+export interface ToolsRequestBuilderGetQueryParameters {
+    /**
+     * Which 'page' of paginated results to return.
+     */
+    page?: number;
+    /**
+     * Number of items returned per page
+     */
+    perPage?: number;
+    /**
+     * Filter tools by toolkit identifier.
+     */
+    toolkitId?: string;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const ToolsRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/tools{?page*,per_page*,toolkit_id*}";
+/**
+ * Mapper for query parameters from symbol name to serialization name represented as a constant.
+ */
+const ToolsRequestBuilderGetQueryParametersMapper: Record<string, string> = {
+    "perPage": "per_page",
+    "toolkitId": "toolkit_id",
+};
+/**
+ * Metadata for all the navigation properties in the request builder.
+ */
+export const ToolsRequestBuilderNavigationMetadata: Record<Exclude<keyof ToolsRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
+    byName: {
+        navigationMetadata: WithNameItemRequestBuilderNavigationMetadata,
+        pathParametersMappings: ["name"],
+    },
+    providers: {
+        requestsMetadata: ProvidersRequestBuilderRequestsMetadata,
+    },
+    toolkits: {
+        requestsMetadata: ToolkitsRequestBuilderRequestsMetadata,
+    },
+};
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const ToolsRequestBuilderRequestsMetadata: RequestsMetadata = {
+    get: {
+        uriTemplate: ToolsRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createList_tools_responseFromDiscriminatorValue,
+        queryParametersMapper: ToolsRequestBuilderGetQueryParametersMapper,
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
