+++ src/dots/v2/actionGateway/tools/item/definition/index.ts
@@ -1,0 +1,74 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createErrorEscapedFromDiscriminatorValue, createTool_definitionFromDiscriminatorValue, type ErrorEscaped, type Tool_definition } from '../../../../../models/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/tools/{name}/definition
+ */
+export interface DefinitionRequestBuilder extends BaseRequestBuilder<DefinitionRequestBuilder> {
+    /**
+     * Retrieves the executable definition for an active Action Gateway tool.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Tool_definition>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 404 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<DefinitionRequestBuilderGetQueryParameters> | undefined) : Promise<Tool_definition | undefined>;
+    /**
+     * Retrieves the executable definition for an active Action Gateway tool.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<DefinitionRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
+}
+/**
+ * Retrieves the executable definition for an active Action Gateway tool.
+ */
+export interface DefinitionRequestBuilderGetQueryParameters {
+    /**
+     * The toolkit identifier used to disambiguate a bare tool name.
+     */
+    toolkitId?: string;
+    /**
+     * The tool version. Omit to retrieve the current version.
+     */
+    version?: string;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const DefinitionRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/tools/{name}/definition{?toolkit_id*,version*}";
+/**
+ * Mapper for query parameters from symbol name to serialization name represented as a constant.
+ */
+const DefinitionRequestBuilderGetQueryParametersMapper: Record<string, string> = {
+    "toolkitId": "toolkit_id",
+};
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const DefinitionRequestBuilderRequestsMetadata: RequestsMetadata = {
+    get: {
+        uriTemplate: DefinitionRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createTool_definitionFromDiscriminatorValue,
+        queryParametersMapper: DefinitionRequestBuilderGetQueryParametersMapper,
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
