+++ src/dots/v2/actionGateway/users/item/index.ts
@@ -1,0 +1,54 @@
+/* tslint:disable */
+/* eslint-disable */
+// Generated by Microsoft Kiota
+// @ts-ignore
+import { createErrorEscapedFromDiscriminatorValue, createGet_user_responseFromDiscriminatorValue, type ErrorEscaped, type Get_user_response } from '../../../../models/index.js';
+// @ts-ignore
+import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
+
+/**
+ * Builds and executes requests for operations under /v2/action-gateway/users/{user_id}
+ */
+export interface WithUser_ItemRequestBuilder extends BaseRequestBuilder<WithUser_ItemRequestBuilder> {
+    /**
+     * Retrieves a derived end-user view containing its sessions and OAuth connections.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {Promise<Get_user_response>}
+     * @throws {ErrorEscaped} error when the service returns a 401 status code
+     * @throws {ErrorEscaped} error when the service returns a 404 status code
+     * @throws {ErrorEscaped} error when the service returns a 429 status code
+     * @throws {ErrorEscaped} error when the service returns a 500 status code
+     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
+     */
+     get(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Get_user_response | undefined>;
+    /**
+     * Retrieves a derived end-user view containing its sessions and OAuth connections.
+     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
+     * @returns {RequestInformation}
+     */
+     toGetRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
+}
+/**
+ * Uri template for the request builder.
+ */
+export const WithUser_ItemRequestBuilderUriTemplate = "{+baseurl}/v2/action-gateway/users/{user_id}";
+/**
+ * Metadata for all the requests in the request builder.
+ */
+export const WithUser_ItemRequestBuilderRequestsMetadata: RequestsMetadata = {
+    get: {
+        uriTemplate: WithUser_ItemRequestBuilderUriTemplate,
+        responseBodyContentType: "application/json",
+        errorMappings: {
+            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
+        },
+        adapterMethodName: "send",
+        responseBodyFactory:  createGet_user_responseFromDiscriminatorValue,
+    },
+};
+/* tslint:enable */
+/* eslint-enable */
