--- src/dots/v2/connections/index.ts
@@ -1,148 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createCreate_connection_responseFromDiscriminatorValue, createErrorEscapedFromDiscriminatorValue, createList_connections_responseFromDiscriminatorValue, serializeCreate_connection_request, serializeCreate_connection_response, type Create_connection_request, type Create_connection_response, type ErrorEscaped, type List_connections_response } from '../../models/index.js';
-// @ts-ignore
-import { ConnectionsItemRequestBuilderRequestsMetadata, type ConnectionsItemRequestBuilder } from './item/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/connections
- */
-export interface ConnectionsRequestBuilder extends BaseRequestBuilder<ConnectionsRequestBuilder> {
-    /**
-     * Gets an item from the ApiSdk.v2.connections.item collection
-     * @param id The connection UUID.
-     * @returns {ConnectionsItemRequestBuilder}
-     */
-     byId(id: string) : ConnectionsItemRequestBuilder;
-    /**
-     * Lists OAuth connections owned by the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<List_connections_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     get(requestConfiguration?: RequestConfiguration<ConnectionsRequestBuilderGetQueryParameters> | undefined) : Promise<List_connections_response | undefined>;
-    /**
-     * Creates or begins authorization for an OAuth connection to an Action Gateway provider.
-     * @param body The request body
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<Create_connection_response>}
-     * @throws {ErrorEscaped} error when the service returns a 400 status code
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 409 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     post(body: Create_connection_request, requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Create_connection_response | undefined>;
-    /**
-     * Lists OAuth connections owned by the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toGetRequestInformation(requestConfiguration?: RequestConfiguration<ConnectionsRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
-    /**
-     * Creates or begins authorization for an OAuth connection to an Action Gateway provider.
-     * @param body The request body
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toPostRequestInformation(body: Create_connection_request, requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
-}
-/**
- * Lists OAuth connections owned by the authenticated team.
- */
-export interface ConnectionsRequestBuilderGetQueryParameters {
-    /**
-     * Which 'page' of paginated results to return.
-     */
-    page?: number;
-    /**
-     * Number of items returned per page
-     */
-    perPage?: number;
-    /**
-     * Filter by provider name.
-     */
-    provider?: string;
-    /**
-     * Field used to sort results.
-     */
-    sort?: string;
-    /**
-     * Sort direction.
-     */
-    sortDirection?: string;
-    /**
-     * Filter by connection status.
-     */
-    status?: string;
-    /**
-     * Filter by end-user identifier.
-     */
-    userId?: string;
-}
-/**
- * Uri template for the request builder.
- */
-export const ConnectionsRequestBuilderUriTemplate = "{+baseurl}/v2/connections{?page*,per_page*,provider*,sort*,sort_direction*,status*,user_id*}";
-/**
- * Mapper for query parameters from symbol name to serialization name represented as a constant.
- */
-const ConnectionsRequestBuilderGetQueryParametersMapper: Record<string, string> = {
-    "perPage": "per_page",
-    "sortDirection": "sort_direction",
-    "userId": "user_id",
-};
-/**
- * Metadata for all the navigation properties in the request builder.
- */
-export const ConnectionsRequestBuilderNavigationMetadata: Record<Exclude<keyof ConnectionsRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
-    byId: {
-        requestsMetadata: ConnectionsItemRequestBuilderRequestsMetadata,
-        pathParametersMappings: ["id"],
-    },
-};
-/**
- * Metadata for all the requests in the request builder.
- */
-export const ConnectionsRequestBuilderRequestsMetadata: RequestsMetadata = {
-    get: {
-        uriTemplate: ConnectionsRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createList_connections_responseFromDiscriminatorValue,
-        queryParametersMapper: ConnectionsRequestBuilderGetQueryParametersMapper,
-    },
-    post: {
-        uriTemplate: ConnectionsRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            400: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            409: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createCreate_connection_responseFromDiscriminatorValue,
-        requestBodyContentType: "application/json",
-        requestBodySerializer: serializeCreate_connection_request,
-        requestInformationContentSetMethod: "setContentFromParsable",
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
