--- src/dots/v2/sessions/item/index.ts
@@ -1,54 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createDelete_session_responseFromDiscriminatorValue, createErrorEscapedFromDiscriminatorValue, type Delete_session_response, type ErrorEscaped } from '../../../models/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/sessions/{session_urn}
- */
-export interface WithSession_urnItemRequestBuilder extends BaseRequestBuilder<WithSession_urnItemRequestBuilder> {
-    /**
-     * Deletes an Action Gateway session owned by the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<Delete_session_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 404 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     delete(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Delete_session_response | undefined>;
-    /**
-     * Deletes an Action Gateway session owned by the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toDeleteRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
-}
-/**
- * Uri template for the request builder.
- */
-export const WithSession_urnItemRequestBuilderUriTemplate = "{+baseurl}/v2/sessions/{session_urn}";
-/**
- * Metadata for all the requests in the request builder.
- */
-export const WithSession_urnItemRequestBuilderRequestsMetadata: RequestsMetadata = {
-    delete: {
-        uriTemplate: WithSession_urnItemRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createDelete_session_responseFromDiscriminatorValue,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
