--- src/dots/v2/toolbelts/item/index.ts
@@ -1,107 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createErrorEscapedFromDiscriminatorValue, createToolbelt_responseFromDiscriminatorValue, type ErrorEscaped, type Toolbelt_response } from '../../../models/index.js';
-// @ts-ignore
-import { ToolsRequestBuilderNavigationMetadata, type ToolsRequestBuilder } from './tools/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/toolbelts/{name}
- */
-export interface WithNameItemRequestBuilder extends BaseRequestBuilder<WithNameItemRequestBuilder> {
-    /**
-     * The tools property
-     */
-    get tools(): ToolsRequestBuilder;
-    /**
-     * Deprecates the latest active version of a toolbelt.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<Toolbelt_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 404 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     delete(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Toolbelt_response | undefined>;
-    /**
-     * Retrieves the latest active version or a specified immutable version of a toolbelt.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<Toolbelt_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 404 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     get(requestConfiguration?: RequestConfiguration<WithNameItemRequestBuilderGetQueryParameters> | undefined) : Promise<Toolbelt_response | undefined>;
-    /**
-     * Deprecates the latest active version of a toolbelt.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toDeleteRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
-    /**
-     * Retrieves the latest active version or a specified immutable version of a toolbelt.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toGetRequestInformation(requestConfiguration?: RequestConfiguration<WithNameItemRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
-}
-/**
- * Retrieves the latest active version or a specified immutable version of a toolbelt.
- */
-export interface WithNameItemRequestBuilderGetQueryParameters {
-    /**
-     * An immutable numeric toolbelt version. Omit to retrieve the latest active version.
-     */
-    version?: string;
-}
-/**
- * Uri template for the request builder.
- */
-export const WithNameItemRequestBuilderUriTemplate = "{+baseurl}/v2/toolbelts/{name}{?version*}";
-/**
- * Metadata for all the navigation properties in the request builder.
- */
-export const WithNameItemRequestBuilderNavigationMetadata: Record<Exclude<keyof WithNameItemRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
-    tools: {
-        navigationMetadata: ToolsRequestBuilderNavigationMetadata,
-    },
-};
-/**
- * Metadata for all the requests in the request builder.
- */
-export const WithNameItemRequestBuilderRequestsMetadata: RequestsMetadata = {
-    delete: {
-        uriTemplate: WithNameItemRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createToolbelt_responseFromDiscriminatorValue,
-    },
-    get: {
-        uriTemplate: WithNameItemRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createToolbelt_responseFromDiscriminatorValue,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
