--- src/dots/v2/toolbelts/item/tools/add/index.ts
@@ -1,61 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createErrorEscapedFromDiscriminatorValue, createToolbelt_responseFromDiscriminatorValue, serializeToolbelt_response, serializeToolbelt_tools, type ErrorEscaped, type Toolbelt_response, type Toolbelt_tools } from '../../../../../models/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/toolbelts/{name}/tools/add
- */
-export interface AddRequestBuilder extends BaseRequestBuilder<AddRequestBuilder> {
-    /**
-     * Adds provider-qualified tool names and creates a new immutable toolbelt version.
-     * @param body The request body
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<Toolbelt_response>}
-     * @throws {ErrorEscaped} error when the service returns a 400 status code
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 404 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     post(body: Toolbelt_tools, requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<Toolbelt_response | undefined>;
-    /**
-     * Adds provider-qualified tool names and creates a new immutable toolbelt version.
-     * @param body The request body
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toPostRequestInformation(body: Toolbelt_tools, requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
-}
-/**
- * Uri template for the request builder.
- */
-export const AddRequestBuilderUriTemplate = "{+baseurl}/v2/toolbelts/{name}/tools/add";
-/**
- * Metadata for all the requests in the request builder.
- */
-export const AddRequestBuilderRequestsMetadata: RequestsMetadata = {
-    post: {
-        uriTemplate: AddRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            400: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            404: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createToolbelt_responseFromDiscriminatorValue,
-        requestBodyContentType: "application/json",
-        requestBodySerializer: serializeToolbelt_tools,
-        requestInformationContentSetMethod: "setContentFromParsable",
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
