--- src/dots/v2/toolbelts/item/tools/index.ts
@@ -1,40 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { AddRequestBuilderRequestsMetadata, type AddRequestBuilder } from './add/index.js';
-// @ts-ignore
-import { RemoveRequestBuilderRequestsMetadata, type RemoveRequestBuilder } from './remove/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/toolbelts/{name}/tools
- */
-export interface ToolsRequestBuilder extends BaseRequestBuilder<ToolsRequestBuilder> {
-    /**
-     * The add property
-     */
-    get add(): AddRequestBuilder;
-    /**
-     * The remove property
-     */
-    get remove(): RemoveRequestBuilder;
-}
-/**
- * Uri template for the request builder.
- */
-export const ToolsRequestBuilderUriTemplate = "{+baseurl}/v2/toolbelts/{name}/tools";
-/**
- * Metadata for all the navigation properties in the request builder.
- */
-export const ToolsRequestBuilderNavigationMetadata: Record<Exclude<keyof ToolsRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
-    add: {
-        requestsMetadata: AddRequestBuilderRequestsMetadata,
-    },
-    remove: {
-        requestsMetadata: RemoveRequestBuilderRequestsMetadata,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
