--- src/dots/v2/tools/item/index.ts
@@ -1,31 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { DefinitionRequestBuilderRequestsMetadata, type DefinitionRequestBuilder } from './definition/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/tools/{name}
- */
-export interface WithNameItemRequestBuilder extends BaseRequestBuilder<WithNameItemRequestBuilder> {
-    /**
-     * The definition property
-     */
-    get definition(): DefinitionRequestBuilder;
-}
-/**
- * Uri template for the request builder.
- */
-export const WithNameItemRequestBuilderUriTemplate = "{+baseurl}/v2/tools/{name}";
-/**
- * Metadata for all the navigation properties in the request builder.
- */
-export const WithNameItemRequestBuilderNavigationMetadata: Record<Exclude<keyof WithNameItemRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
-    definition: {
-        requestsMetadata: DefinitionRequestBuilderRequestsMetadata,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
