--- src/dots/v2/tools/providers/index.ts
@@ -1,52 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createErrorEscapedFromDiscriminatorValue, createList_providers_responseFromDiscriminatorValue, type ErrorEscaped, type List_providers_response } from '../../../models/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/tools/providers
- */
-export interface ProvidersRequestBuilder extends BaseRequestBuilder<ProvidersRequestBuilder> {
-    /**
-     * Lists Action Gateway providers and their connection requirements.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<List_providers_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     get(requestConfiguration?: RequestConfiguration<object> | undefined) : Promise<List_providers_response | undefined>;
-    /**
-     * Lists Action Gateway providers and their connection requirements.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toGetRequestInformation(requestConfiguration?: RequestConfiguration<object> | undefined) : RequestInformation;
-}
-/**
- * Uri template for the request builder.
- */
-export const ProvidersRequestBuilderUriTemplate = "{+baseurl}/v2/tools/providers";
-/**
- * Metadata for all the requests in the request builder.
- */
-export const ProvidersRequestBuilderRequestsMetadata: RequestsMetadata = {
-    get: {
-        uriTemplate: ProvidersRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createList_providers_responseFromDiscriminatorValue,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
