--- src/dots/v2/users/index.ts
@@ -1,89 +1,0 @@
-/* tslint:disable */
-/* eslint-disable */
-// Generated by Microsoft Kiota
-// @ts-ignore
-import { createErrorEscapedFromDiscriminatorValue, createList_users_responseFromDiscriminatorValue, type ErrorEscaped, type List_users_response } from '../../models/index.js';
-// @ts-ignore
-import { type WithUser_ItemRequestBuilder, WithUser_ItemRequestBuilderRequestsMetadata } from './item/index.js';
-// @ts-ignore
-import { type BaseRequestBuilder, type KeysToExcludeForNavigationMetadata, type NavigationMetadata, type Parsable, type ParsableFactory, type RequestConfiguration, type RequestInformation, type RequestsMetadata } from '@microsoft/kiota-abstractions';
-
-/**
- * Builds and executes requests for operations under /v2/users
- */
-export interface UsersRequestBuilder extends BaseRequestBuilder<UsersRequestBuilder> {
-    /**
-     * Gets an item from the ApiSdk.v2.users.item collection
-     * @param user_id The end-user identifier.
-     * @returns {WithUser_ItemRequestBuilder}
-     */
-     byUser_id(user_id: string) : WithUser_ItemRequestBuilder;
-    /**
-     * Lists end-user identifiers derived from sessions and OAuth connections for the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {Promise<List_users_response>}
-     * @throws {ErrorEscaped} error when the service returns a 401 status code
-     * @throws {ErrorEscaped} error when the service returns a 429 status code
-     * @throws {ErrorEscaped} error when the service returns a 500 status code
-     * @throws {ErrorEscaped} error when the service returns a 4XX or 5XX status code
-     */
-     get(requestConfiguration?: RequestConfiguration<UsersRequestBuilderGetQueryParameters> | undefined) : Promise<List_users_response | undefined>;
-    /**
-     * Lists end-user identifiers derived from sessions and OAuth connections for the authenticated team.
-     * @param requestConfiguration Configuration for the request such as headers, query parameters, and middleware options.
-     * @returns {RequestInformation}
-     */
-     toGetRequestInformation(requestConfiguration?: RequestConfiguration<UsersRequestBuilderGetQueryParameters> | undefined) : RequestInformation;
-}
-/**
- * Lists end-user identifiers derived from sessions and OAuth connections for the authenticated team.
- */
-export interface UsersRequestBuilderGetQueryParameters {
-    /**
-     * Which 'page' of paginated results to return.
-     */
-    page?: number;
-    /**
-     * Number of items returned per page
-     */
-    perPage?: number;
-}
-/**
- * Uri template for the request builder.
- */
-export const UsersRequestBuilderUriTemplate = "{+baseurl}/v2/users{?page*,per_page*}";
-/**
- * Mapper for query parameters from symbol name to serialization name represented as a constant.
- */
-const UsersRequestBuilderGetQueryParametersMapper: Record<string, string> = {
-    "perPage": "per_page",
-};
-/**
- * Metadata for all the navigation properties in the request builder.
- */
-export const UsersRequestBuilderNavigationMetadata: Record<Exclude<keyof UsersRequestBuilder, KeysToExcludeForNavigationMetadata>, NavigationMetadata> = {
-    byUser_id: {
-        requestsMetadata: WithUser_ItemRequestBuilderRequestsMetadata,
-        pathParametersMappings: ["user_id"],
-    },
-};
-/**
- * Metadata for all the requests in the request builder.
- */
-export const UsersRequestBuilderRequestsMetadata: RequestsMetadata = {
-    get: {
-        uriTemplate: UsersRequestBuilderUriTemplate,
-        responseBodyContentType: "application/json",
-        errorMappings: {
-            401: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            429: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            500: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-            XXX: createErrorEscapedFromDiscriminatorValue as ParsableFactory<Parsable>,
-        },
-        adapterMethodName: "send",
-        responseBodyFactory:  createList_users_responseFromDiscriminatorValue,
-        queryParametersMapper: UsersRequestBuilderGetQueryParametersMapper,
-    },
-};
-/* tslint:enable */
-/* eslint-enable */
